+++ R/theme.R
@@ -35,8 +35,8 @@ set_theme <- function(name = c("billboard", "insight", "graph")) {
 theme_dependency <- function() {
   htmlDependency(
     name = "billboard-theme", 
-    version = "1.9.5", 
-    src = "htmlwidgets/lib/billboard-1.9",
+    version = "1.10.0", 
+    src = "htmlwidgets/lib/billboard-1.10",
     package = "billboarder",
     stylesheet = getOption(
       x = "billboard.theme",